Week 7 mostly working
I can't get the editor interface quite right yet. I reverted to the more
efficient (I think) method of checking if we are over bounds in dropJ
and takeJ instead of delegating the work to +++.

I also corrected an error that was reversing the input. Still does not
load a file correctly.

When I browse the example buffer, I can see lines at negative indices,
but am unable to visit them.